Css de düzenlemeler yapıldı ve order list çekilde buton fonksiyonelliği eklenecek
