+++ HammalWeb/Areas/Customer/Views/Category/CategoryDetail.cshtml
@@ -30,10 +30,10 @@
               <div class="col-md-3">
                 <div class="d-flex justify-content-between">
                    
-                  <div id="wrapper">
-                    <a class="my-super-cool-btn ml-3" style="max-width:50%" onclick="window.location.href='@Url.Action("ServiceClaimView","Service",new {area="Customer",altCategoryId=altCategory.Id})'">
-                      <span>""</span>
-                    </a>
+                  <div class="col-md-12">
+                    <button class="btn btn-warning" style="max-width:50%" onclick="window.location.href='@Url.Action("ServiceClaimView","Service",new {area="Customer",altCategoryId=altCategory.Id})'">
+                      <span>Teklif Al</span>
+                    </button>
                   </div>
                     
                   </div>
