+++ HammalWeb/Areas/Customer/Views/Service/EmployeeForm.cshtml
@@ -10,16 +10,20 @@
       <hr />
 
       <div class="form-floating mb-3 col-md-6">
+        <span><b>Kategori</b></span>
         <select class="form-control" id="categoryDropdown"></select>
       </div>
       <div class="form-floating mb-3 col-md-6">
+        <span><b>Alt Kategori</b></span>
         <select class="form-control" id="altCategoryDropdown"></select>
       </div>
-      <div class="col-md-6">
-        <input type="number" asp-for="@Model.Price" />
+      <div class="form-floating mb-3 col-md-6">
+        <span><b>Ücret</b></span>
+        <input class="form-control" type="number" asp-for="@Model.Price" />
       </div>
-      <div class="col-md-6">
-        <input type="text" asp-for="@Model.Abilities" />
+      <div class="form-floating mb-3 col-md-6">
+        <span><b>Yetenekler</b></span>
+        <input class="form-control" type="text" asp-for="@Model.Abilities" />
       </div>
    
       <button type="submit" class="w-100 btn btn-lg btn-primary">Register</button>
