+++ HammalWeb/Views/Shared/_Footer.cshtml
@@ -1,0 +1,31 @@
+﻿
+  <div class="container mt-3">
+    <div class="row">
+      <div class="col-sm-6 col-md-3 item">
+        <h3>Hizmetler</h3>
+        <ul>
+          <li><a href="#">Web tasarımı</a></li>
+          <li><a href="#">Geliştirme</a></li>
+          <li><a href="#">Barındırma</a></li>
+        </ul>
+      </div>
+      <div class="col-sm-6 col-md-3 item">
+        <h3>Hakkında</h3>
+        <ul>
+          <li><a href="#">Şirket</a></li>
+          <li><a href="#">Ekip</a></li>
+          <li><a href="#">Kariyer</a></li>
+        </ul>
+      </div>
+      <div class="col-md-6 item text">
+        <h3>Şirket Adı</h3>
+        <p>Praesent sed lobortis mi. Suspendisse vel placerat ligula. Vivamus ac sem lacus. Ut vehicula rhoncus elementum. Etiam quis tristique lectus. Aliquam in arcu eget velit pulvinar dictum vel in justo.</p>
+      </div>
+      <div class="col item social"><a href="#"><i class="bi bi-facebook"></i></a><a href="#"><i class="bi bi-twitter"></i></a><a href="#"><i class="bi bi-snapchat"></i></a><a href="#"><i class="bi bi-instagram"></i></a></div>
+    </div>
+  </div>
+  <div class="container-fluid bg-dark">
+    <div class="container">
+      <p class="text-center text-white m-0 py-3">Şirket Adı © 2018</p>
+    </div>
+  </div>
