+++ HammalWeb/Views/Shared/_Notification.cshtml
@@ -17,3 +17,4 @@
         toastr.error('@TempData["error"]')
     </script>
 }
+
