+++ HammalWeb/wwwroot/css/site.css
@@ -27,7 +27,7 @@ footer {
     width: 100%;
     height: auto;
 }
-
+/*
 #wrapper {
   width: 100%;
   height: 100%;
@@ -250,3 +250,4 @@ footer {
     opacity: 0;
   }
 }
+*/
